apps/discord/requirements.txt
# Discord Integration Requirements
# Python 3.14.2 Required
# Updated: January 11, 2026

# Discord Bot Framework
discord.py>=2.4.0
PyNaCl>=1.5.0

# Audio Processing
pydub>=0.25.1
av>=13.1.0

# Async Support
python-dotenv>=1.0.1
websockets>=14.1
python-dateutil>=2.9.0.post0

# Networking & HTTP
requests>=2.32.3
urllib3>=2.3.0

# Configuration
pyyaml>=6.0.2
pydantic>=2.10.3

# Scheduling
apscheduler>=3.11.0

# System Utilities
psutil>=6.1.0
colorama>=0.4.6
loguru>=0.7.3
